This change is user visible.
Allwoing to mix Enum, int, and str makes it hard to do the
compiled code to do the same thing as the interpreted one.

I believe it is better for readability of the user code,
the compiler and basically everything between to not allow
that. Only int and enum types shall be used.

This clarifies the situation.